requirements.txt
Django==5.0.3
psycopg2-binary==2.9.9
django-environ==0.11.2
djangorestframework==3.15.1
django-cors-headers